==> ImpactDataMining/Hurricane_Ian/02_Code/01_input.ipynb <==
This Notebook is designed to set up all essential inputs for the project, encompassing folder paths, keywords, model parameters, and more. The input_path, which represents the location of this file, is crucial to be pasted at the start of every Python file in this project. Doing so ensures that all required inputs are correctly loaded, thereby enabling the successful execution of the code.

Recommended Google Colab Runtime Type: CPU (default).

# Specify the directory path where the output of this input file will be saved
input_path =  "/content/drive/My Drive/ImpactDataMining/Hurricane_Ian/01_Input"

# Specify the file path and name for the input document
document_path = "/content/drive/My Drive/ImpactDataMining/Hurricane_Ian/01_Input"
document_name = "Hurricane Ian_PVRR.docx"

# Specify the directory path where all output generated by the notebooks will be stored
output_path =  "/content/drive/My Drive/ImpactDataMining/Hurricane_Ian/03_Output"

# Input for 1_preprocessing.ipynb
# Specify headings to exclude from the table of contents in the optional file (00_optional_customize_scope_by_toc.ipynb), if applicable
drop_headings = [
    'PREFACE',
    'ATTRIBUTION GUIDANCE',
    'ACKNOWLEDGMENTS',
    'TABLE OF CONTENTS',
    'Recommended Response Strategy',
    'Appendix A. Loss Estimates',
    'Appendix B. Evacuation Orders',
    'Appendix C. Hazard Observations',
    'References',
    ]

# Input for 2a_extracting_impact_sentences.ipynb and 2b_selecting_threshold.ipynb
# Define keywords based on categories and combine at the end
building_types = [
    'family housing damage', 'school damage', 'mobile home damage', 'manufactured home damage', 'commercial facility damage', 'hospital damage',
    'religious infrastructure damage', 'historical building damage', 'industrial facility damage', 'agricultural infrastructure damage',
    'fire station damage', 'police station damage', 'parking damage', 'critical facility damage', 'shelter damage'
    ]
building_components = [
    'roof damage', 'secondary water resistance damage', 'roof deck damage', 'roof-wall connection damage', 'toenail damage',
    'hurricane strap damage', 'shutter damage', 'roof cover damage', 'shingle damage', 'built-up roofing damage',
    'single-ply membrane damage', 'wall damage', 'roof frame damage', 'wood truss damage', 'open web steel joist damage', 'joist spacing',
    'window damage', 'tie-downs damage', 'foundation anchorage damage', 'door damage', 'glazing damage', 'cladding damage'
]
other_infrastructure = [
    'airport damage', 'port damage', 'bridge damage', 'road damage', 'causeway damage', 'pavement damage', 'seawall damage', 'levee damage',
    'water tank damage', 'power plants damage', 'offshore structure damage'
]
structure_related_terms = [
     'structural behavior', 'structural damage', 'structural failure mechanisms', 'infrastructure resilience', 'structural connection damage',
     'building interior damage', 'building exterior damage', 'building envelope damage', 'building damage', 'lateral force resisting systems',
     'undamaged infrastructure'
]
terrestrial_damage = [
    'erosion', 'washout', 'wind-borne debris', 'scour', 'tree-related damage'
]
community_damage = [
    'fatalities', 'economic loss', 'power outage', 'healthcare disruption', 'agricultural damage', 'uninsured loss',
    'telecommunication system disruption', 'transportation disruption', 'utility outage', 'water outage', 'service loss', 'supply halt',
    'temporary medical facility', 'transaction suspension'
]
non_impact_terms = [
    'storm tracking', 'building code', 'research support'
    ]

keywords_struct = building_types + building_components + other_infrastructure + structure_related_terms + terrestrial_damage
keywords_comm = community_damage
keywords_non_impact = non_impact_terms

keywords = keywords_struct + keywords_comm + keywords_non_impact

# Input for 2a_extracting_impact_sentences.ipynb and 2b_selecting_threshold.ipynb
batch_size = 8 # number of sentences / batch for BART large MNLI
threshold = .9 # entailment threshold for extracting impact sentences
overlap_tokens =200 # overlapping text between summary chunk for BART large CNN

The following section compiles all previously defined inputs and saves them. Therefore, there is no need to make any edits beyond this point.

# Combine all the inputs
input = {
    'document_path': document_path,
    'output_path': output_path,
    'document_name': document_name,
    'drop_headings': drop_headings,
    'keywords_struct': keywords_struct,
    'keywords_comm': keywords_comm,
    'keywords_non_impact': keywords_non_impact,
    'keywords': keywords,
    'batch_size': batch_size,
    'threshold': threshold,
    'overlap_tokens': overlap_tokens
    }

import os
import json

from google.colab import drive

def current_path():
  print("Current working directory")
  print(os.getcwd())
  print()

current_path()
drive.mount('/content/drive')
os.chdir(input_path)
current_path()

# Convert the Python dictionary to a JSON string
input = json.dumps(input, indent=4)

with open('0_input.json', 'w') as file:
    file.write(input)
